// backend/src/common/types/common.types.ts
export enum SyncTabla {
  PRODUCTOS = 'productos',
  LLENADORAS = 'llenadoras',
  CABEZALES = 'cabezales',
  EAN = 'ean',
  PRESENTACIONES = 'presentaciones',
  ROLES = 'roles',
  USUARIOS = 'usuarios',
  MOTIVO_BAJAS = 'motivo_bajas',
  NUMERO_BIDON = 'numero_bidon',
  PRODUCCION = 'produccion',
  ASOCIACION_PRODUCCION = 'asociacion_produccion',
  PESOS = 'pesos',
  SINCRONISMO = 'sincronismo',
}

export enum CabezalID {
  A = 'A',
  B = 'B',
  C = 'C',
}

export enum UserRole {
  ADMIN = 'ADMINISTRADOR',
  USER = 'USUARIO',
  SUPERVISOR = 'SUPERVISOR',
  CALIDAD = 'CALIDAD',
  MANTENIMIENTO = 'MANTENIMIENTO',
}